requirements.txt
Pillow>=5.3.0
PyGithub
aiofiles
aiohttp
anime_downloader
async_generator
asyncurban
beautifulsoup4
bs4
bwb==3.0.0
chromedriver-py==88.0.4324.96
cryptography==3.1.1
cfscrape
covid
cowpy
cryptg
dnspython
emoji
gTTS-token>=1.1.3
gTTS>=2.0.1
geopy
gitpython
google-api-python-client==1.8.0
google-auth-oauthlib
google_images_download>=2.7.1
googletrans==3.1.0a0
gsearch
hachoir
httplib2<=0.19.0
humanize
lxml
lyricsgenius
oauth2client
psycopg2-binary
pySmartDL
pybase64>=0.4.0
pyfiglet
pylast
pymongo
python-barcode
python-dotenv
python-magic
pytube
pytz
qrcode
regex
requests>=2.18.4
search-engine-parser>=0.4.2
selenium
speedtest-cli
sqlalchemy==1.3.23
telegraph
telethon>=1.24.0
tswift
urbandict
validators
wheel
wikipedia>=1.4.0
youtube-dl>=2021.1.8
youtube-search-python>=1.3.2
youtube-search>=1.1.1
pygments
catsart
tswift